Requested changes by Zodern
