+++ package.json
@@ -37,7 +37,7 @@
     "lint:docs-configs": "node tests/validate-docs.js",
     "lint:css": "cd docs && stylelint css/*.css",
     "docs": "cd docs && bundle exec jekyll serve",
-    "postinstall": "opencollective-postinstall",
+    "postinstall": "opencollective-postinstall || (exit 0)",
     "clean-shrinkwrap": "cd tests && node clean-shrinkwrap.js"
   },
   "devDependencies": {
