+++ src/plugins/meteor/index.js
@@ -1,4 +1,10 @@
 import * as _commands from './commands';
+import {
+  configHasMailUrl,
+  configHasMongoUrl,
+  normalizeMailUrl,
+  normalizeMongoUrl
+} from './utils';
 import _validator from './validate';
 import { defaultsDeep } from 'lodash';
 import traverse from 'traverse';
@@ -38,6 +44,18 @@ export function prepareConfig(config) {
   // and meteorhacks/meteord, but they allow the PORT env
   // variable to override it.
   config.app.docker.imagePort = config.app.docker.imagePort || 3000;
+
+  if (configHasMailUrl(config)) {
+    const mailUrl = config.app.env.MAIL_URL;
+
+    config.app.env.MAIL_URL = normalizeMailUrl(mailUrl);
+  }
+
+  if (configHasMongoUrl(config)) {
+    const mongoUrl = config.app.env.MONGO_URL;
+
+    config.app.env.MONGO_URL = normalizeMongoUrl(mongoUrl);
+  }
 
   return config;
 }
