+++ src/utils.js
@@ -286,55 +286,3 @@ export function filterArgv(argvArray, argv, unwanted) {
 
   return result;
 }
-
-export function configHasMailUrl(config) {
-  return config && config.app && config.app.env && config.app.env.MAIL_URL;
-}
-
-export function configHasMongoUrl(config) {
-  return config && config.app && config.app.env && config.app.env.MONGO_URL;
-}
-
-export function normalizeMailUrl(mailUrl) {
-  let url = mailUrl;
-
-  // Regex get everything what's between smtp|smtps:// and the last @
-  const re = /^(smtp|smtps):\/\/(.*)@/g.exec(url);
-
-  if (re && re.length > 0) {
-    const arr = re[2].split(':');
-
-    // Username is returned by .shift() if needed
-    arr.shift();
-
-    // Use arr.join('') because user might have more that one : in the pass
-    const fullpassword = arr.join('');
-    const encodedPassword = encodeURIComponent(fullpassword);
-
-    url = url.replace(fullpassword, encodedPassword);
-  }
-
-  return url;
-}
-
-export function normalizeMongoUrl(mongoUrl) {
-  let url = mongoUrl;
-
-  // Regex get everything what's between mongodb:// and the last @
-  const re = /mongodb:\/\/(.*)@/g.exec(url);
-
-  if (re && re.length > 0) {
-    const arr = re[1].split(':');
-
-    // Username is returned by .shift() if needed
-    arr.shift();
-
-    // Use arr.join('') because user might have more that one : in the pass
-    const fullpassword = arr.join('');
-    const encodedPassword = encodeURIComponent(fullpassword);
-
-    url = url.replace(fullpassword, encodedPassword);
-  }
-
-  return url;
-}
